+++ platform/src/app/(main)/(primary)/yomu/layout.tsx
@@ -1,9 +1,9 @@
 // Layout to apply yomu color scheming
-import '@/styles/platform/apps/yomu.scss';
+import '@/styles/pages/yomu/_layout.scss';
 
 export default function YomuLayout({children} : {children: React.ReactNode}) {
     return (
-        <div id="yomu-app">
+        <div id="yomu-layout">
             {children}
         </div>
     )