+++ platform/src/app/(main)/(secondary)/kumitateru/layout.tsx
@@ -1,9 +1,9 @@
 // Layout to apply kumitateru color scheming
-import '@/styles/platform/apps/kumitateru.scss';
+import '@/styles/pages/kumitateru/_layout.scss'
 
 export default function KumitateruLayout({children} : {children: React.ReactNode}) {
     return (
-        <div id="kumitateru-app">
+        <div id="kumitateru-layout">
             {children}
         </div>
     )