+++ platform/src/app/(main)/(secondary)/kumitateru/partDialogue.tsx
@@ -1,7 +1,7 @@
 import { Dialog, DialogContent, DialogTitle } from "@mui/material"
 
-import '@/styles/platform/components/partDetailDialog.scss';
-import InfoItem from "@/components/platform/infoItem";
+import '@/styles/components/kumitateru/_part-dialog.scss';
+import InfoItem from "@/components/infoItem";
 import { CPU, GPU, Motherboard, PSU, RAM } from "@/types/kumitateru";
 
 export default function PartDetailDialog(